[ kotlin ] Rename package
