+++ test/org/lice/parse/LexCommentTest.kt
@@ -1,4 +1,4 @@
-package org.lice.parse2
+package org.lice.parse
 
 import org.junit.Test
