+++ test/org/lice/parse/LexIdentifierTest.kt
@@ -1,4 +1,4 @@
-package org.lice.parse2
+package org.lice.parse
 
 import org.junit.Test
