+++ test/org/lice/parse/LexNumberTest.kt
@@ -1,4 +1,4 @@
-package org.lice.parse2
+package org.lice.parse
 
 import org.junit.Rule
 import org.junit.Test
