+++ test/org/lice/parse/LexStringTest.kt
@@ -1,4 +1,4 @@
-package org.lice.parse2
+package org.lice.parse
 
 import org.intellij.lang.annotations.Language
 import org.junit.Assert.assertEquals
